requirements.txt
2to3==1.0
alabaster==1.0.0
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asciimatics==1.15.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
Cython==3.0.11
defusedxml==0.7.1
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.112.0
fastapi-cli==0.0.5
fastapi-simple-security==1.3.0
filelock==3.15.4
Flask==3.0.3
flightradar24==0.3.1
fonttools==4.53.1
fsspec==2024.6.1
full==0.0.3.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.3
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Library==0.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
narwhals==1.6.0
networkx==3.3
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
paracelsus==0.7.0
pefile==2024.8.26
pillow==10.4.0
pluggy==1.5.0
protobuf==5.28.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydot==3.0.1
pyfiglet==1.0.2
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.4
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqladmin==0.18.0
SQLAlchemy==2.0.32
sqlalchemy-data-model-visualizer==0.1.3
starlette==0.37.2
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
text-unidecode==1.3
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.5
vidgear==0.3.3
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
WTForms==3.1.2
